ci(Update): Update default value settings in DataBaseContext
The default value settings in the Entity Framework modelBuilder for
TPlace have been updated. The defaults for 'CanBeDeleted' and 'IsOpen'
attributes have been changed to use the HasDefaultValueSql method
instead of HasDefaultValue. This change ensures that the values are set
at the SQL server level rather than at the application level.