ci(Update): Update interaction with Location Management Card in DashBoard
Updated the code to change the mouse down event from the Location
Management UserControl to the card itself in the dashboard. This will
improve the user's interaction experience as they can now click anywhere
on the Location Management Card to navigate to the respective page,
rather than only on the UserControl.