ci(Update): Add AssemblyInfo to SvgToXaml project
A new AssemblyInfo file is introduced in the SvgToXaml project. Within
this file, theme-specific resource dictionary location is set to None,
and the generic resource dictionary location is set to SourceAssembly.
These describes where to find resources if they are not found in the
page, application, or any theme-specific resource dictionaries.