ci(Update): Apply rounded window corner style across application
The commit introduces a new window style preference for rounded corners
in all the windows within the MyExpenses application. It uses the
WindowInteropHelper class to ensure handles are created for each window
and applies the DwmWindowCornerPreference to them. Additionally, the
glow brush attribute has been removed from the
SelectDatabaseFileWindow.xaml file.