ci(Update): Make SaveToCloud method static and improve cloud export
The SaveToCloud method in WelcomePage.xaml.cs has been modified to be
static, enabling it to be invoked without creating an object.
Furthermore, the process of exporting to the cloud has been improved.
Now, each existing database in the selected list gets uploaded to the
"Databases" directory in the cloud storage via the DropboxService. These
changes help organize and facilitate cloud file management.