ci(Update): Update database saving methods to be asynchronous
The methods for saving the database to local storage have been updated
to be asynchronous. Changes have been made across several functions to
implement this, with the use of the 'await' keyword and appropriate
Tasks. This will help to improve the performance of the application as
it will not block the UI thread.