ci(Update): Refactor import methods and remove redundant logs
Removed the logs that informed about successful database save or upload,
as these were redundant. Refactored the methods `ImportFromCloudAsync`
and `ImportFromLocal` to improve logging. Now, the process of starting
the import, downloading or copying the file, and successful completion
of each step is logged. This provides more granular insight into the
import process.