ci(Update): Add localization support to dashboard data grid
A new utility class, LanguageService, has been introduced to convert
System.Globalization.CultureInfo objects into local languages used by
the FilterDataGrid. In the DashBoardPage constructor, the current
culture is converted to a local language and stored as a property. This
local language is then bound to the FilterDataGrid's FilterLanguage
property in the XAML file, enabling localized filtering.