ci(Update): Rename RemoveDatabaseFileResources to SelectDatabaseFileWindowResources
In the WPF application, the resource files and classes related to
RemoveDatabaseFile have been renamed to SelectDatabaseFileWindow. This
includes updates not only to the names of resource files, but also
changes in the application's .csproj and related classes. The change
allows better semantic coherence between file names and their associated
functionality.