ci(Update): Change ToKmlFile method to void
The ToKmlFile method in GoogleEarth class has been modified from
returning a string to void. This change resulted in discarding the file
path as a return value. The file path is now simply given directly as a
parameter. Alongside the change, Path.ChangeExtension to set the file
extension as ".kml" has also been removed.