ci(Update): Replace button with checkbox in AccountValueTrendControl
The control elements in the AccountValueTrendControl, previously using
buttons, have been updated to checkboxes. The purpose of this
modification is to change the way users interact with the visibility of
series in the cartesian chart on the UI. Now, the visibility of each
line series can be toggled using corresponding checkboxes, thus
improving user experience.