ci(Update): Add color configuration to settings window
The code has been modified to allow configuration of primary and
secondary color for the theme in the Settings Window. Reading of these
values has been implemented so that they're directly taken from
AppearanceControl, and an Interface theme is saved with these values.
Consequently, the application loads the interface configuration with the
selected primary and secondary colors.