ci(Update): Add language setting feature in the settings window
This commit introduces a language setting feature in the settings
window. A new converter, "FirstCharUpperConverter," is added to
capitalize the first letters of displayed languages. In the
SettingsWindow.xaml.cs file, languages are fetched from the database
context and added to the "CultureInfos" list for use in language
selection in the UI.