ci(Update): Update binding modes and triggers in LocationManagementPage
The bindings for some menu items and a combobox in
LocationManagementPage.xaml have been updated to the OneWay mode and the
UpdateSourceTrigger has been set to PropertyChanged. This ensures the
User Interface (UI) refreshes immediately when properties change.
Corresponding properties in the LocationManagementPage.xaml.cs file have
been moved to an `UpdateLanguage()` method, which is invoked during
theme changes.