ci(Update): Convert button and data grid strings to DependencyProperties
The properties representing the button titles, DataGrid headers, and
ComboBox hints have been turned into DependencyProperty so that they can
be bound to in XAML. An "UpdateLanguage" method has been added that
updates the language of these properties using the current system
language, improving localization support. Additionally, an event
listener was added for a LanguageChanged event to trigger the language
update.