ci(Update): Update data binding mode across AddEditLocationWindow
The data binding mode for several elements in AddEditLocationWindow.xaml
has been changed to 'OneWay'. This ensures that the UI is updated
whenever the bound property changes in the view model. Furthermore, an
'UpdateSourceTrigger' has been set to 'PropertyChanged' to update the
source immediately whenever the target property changes. This change
improves the responsiveness of the UI and aligns the data binding
strategy across the window.