ci(Update): Refactor language loading and reduce namespace verbosity
Removed unnecessary namespace details in App.xaml.cs to streamline the
code and improved the language loading function. If language is not
specified, the system will now default to either the current user
interface culture, or the default language from the supported languages
in the database. The configuration is then updated with the chosen
language.