ci(Update): Remove MyExpenses.Sql.Test project
The MyExpenses.Sql.Test project, alongside its related test cases and
configuration files, were removed. The removal involved .cs files, the
relevant .csproj and the references to it in the solution file. This
decision was taken either due to code refactoring or the tests being
outdated and no longer needed.