ci(Update): Add MyExpenses.Wpf.CleanRelease project for cleaning builds
This commit introduces a new WPF project named
MyExpenses.Wpf.CleanRelease. The project includes functionality to
select directories and clean up unnecessary files such as .pdb, .xml,
and .json files. Additionally, a minor update is made to the CHANGELOG
to reflect the new translation for the budget header in French
resources.